Move SSLFingerprints into separate header file.
